Move to lowlevel compute motor
TODO : Beware we wait 200 ms before motor_release
