+++ C_Controller/controller.c
@@ -63,71 +63,57 @@ bool sensor_test(float(*sensor)(), float min, float max, float maxstddev)
 
 int self_tests()
 {
+
 //    DEBUG_PRINT("Start self tests");
     int test_bits = 0xFFFFFFFF;
-//
-//    // TODO test 'Arret imminent' ?
-//
-//    printf("Buzzer low\n");
-//    check(&test_bits, 1, buzzer_low      (On )); wait_ms(1000);
-//    check(&test_bits, 1, buzzer_low      (Off)); // start pos
-//
-//    printf("Buzzer medium\n");
-//    check(&test_bits, 1, buzzer_medium      (On )); wait_ms(1000);
-//    check(&test_bits, 1, buzzer_medium      (Off)); // start pos
-//
-//    printf("Buzzer high\n");
-//    check(&test_bits, 1, buzzer_high      (On )); wait_ms(1000);
-//    check(&test_bits, 1, buzzer_high      (Off)); // start pos
-//
-//    printf("Red light\n");
-//    check(&test_bits, 2, light_red   (On )); wait_ms(1000);
-//    check(&test_bits, 2, light_red   (Off)); // start pos
-//
-//    printf("Yellow light\n");
-//    check(&test_bits,  9, light_yellow(On )); wait_ms(1000);
-//    check(&test_bits, 10, light_yellow(Off)); // start pos
-//
-//    printf("Green light\n");
-//    check(&test_bits,  9, light_green(On )); wait_ms(1000);
-//    check(&test_bits, 10, light_green(Off)); // start pos
-//
-//    check(&test_bits, 5, init_sensors());
-//    check(&test_bits, 11, sensors_start());
-//
-//    check(&test_bits, 5, sensor_test(get_sensed_VolM_Lpm  , -100,  100, 2)); printf("Rest    Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm  ());
-//    check(&test_bits, 6, sensor_test(get_sensed_P_cmH2O   ,  -20,  100, 2)); printf("Rest    Paw  cmH2O:%+.1g\n", get_sensed_P_cmH2O   ());
-//    check(&test_bits, 7, sensor_test(get_sensed_Patmo_mbar,  900, 1100, 2)); printf("Rest    Patmo mbar:%+.1g\n", get_sensed_Patmo_mbar());
-//
-//    check(&test_bits, 3, init_valve());
-//    check(&test_bits, 3, valve_exhale());
-//
-//    check(&test_bits, 4, init_motor());
-//    printf("Exhale  Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
-//    check(&test_bits, 4, motor_release());
-//    wait_ms(3000);
-//    check(&test_bits, 4, motor_stop());
-//    printf("Release Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
-//    check(&test_bits, 3, valve_inhale());
-//    printf("Inhale  Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
-//
-//    motor_press_constant(400, 1000);
-//    wait_ms(1000);
-//    motor_stop();
-//
-//    motor_release();
-//    wait_ms(3000);
-//
-//	light_green(On);
-//	valve_inhale();
-//	motor_press_constant(MOTOR_STEP_TIME_US_MIN, 3800);
-//	wait_ms(1000);
-//	motor_stop();
-//	valve_exhale();
-//	motor_release();
-//	light_green(Off);
-//	wait_ms(3000);
 
+#ifndef TEST_MOTOR
+    // TODO test 'Arret imminent' ?
+
+    printf("Buzzer low\n");
+    check(&test_bits, 1, buzzer_low      (On )); wait_ms(1000);
+    check(&test_bits, 1, buzzer_low      (Off)); // start pos
+
+    printf("Buzzer medium\n");
+    check(&test_bits, 1, buzzer_medium      (On )); wait_ms(1000);
+    check(&test_bits, 1, buzzer_medium      (Off)); // start pos
+
+    printf("Buzzer high\n");
+    check(&test_bits, 1, buzzer_high      (On )); wait_ms(1000);
+    check(&test_bits, 1, buzzer_high      (Off)); // start pos
+
+    printf("Red light\n");
+    check(&test_bits, 2, light_red   (On )); wait_ms(1000);
+    check(&test_bits, 2, light_red   (Off)); // start pos
+
+    printf("Yellow light\n");
+    check(&test_bits,  9, light_yellow(On )); wait_ms(1000);
+    check(&test_bits, 10, light_yellow(Off)); // start pos
+
+    printf("Green light\n");
+    check(&test_bits,  9, light_green(On )); wait_ms(1000);
+    check(&test_bits, 10, light_green(Off)); // start pos
+
+    check(&test_bits, 5, init_sensors());
+    check(&test_bits, 11, sensors_start());
+
+    check(&test_bits, 5, sensor_test(get_sensed_VolM_Lpm  , -100,  100, 2)); printf("Rest    Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm  ());
+    check(&test_bits, 6, sensor_test(get_sensed_P_cmH2O   ,  -20,  100, 2)); printf("Rest    Paw  cmH2O:%+.1g\n", get_sensed_P_cmH2O   ());
+    check(&test_bits, 7, sensor_test(get_sensed_Patmo_mbar,  900, 1100, 2)); printf("Rest    Patmo mbar:%+.1g\n", get_sensed_Patmo_mbar());
+
+    check(&test_bits, 3, init_valve());
+    check(&test_bits, 3, valve_exhale());
+
+    check(&test_bits, 4, init_motor());
+    printf("Exhale  Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
+    check(&test_bits, 4, motor_release());
+    wait_ms(3000);
+    check(&test_bits, 4, motor_stop());
+    printf("Release Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
+    check(&test_bits, 3, valve_inhale());
+    printf("Inhale  Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
+    motor_release();
+    wait_ms(3000);
 
     //printf("Press   Pdiff  Lpm:%+.1g\n", get_sensed_VolM_Lpm());
     //check(&test_bits, 4, motor_stop());
@@ -141,6 +127,7 @@ int self_tests()
     // TODO check(&test_bits, 8, motor_pep_...
     wait_ms(10000);*/
 
+#else
 	init_motor();
 	init_sensors();
 	sensors_start();
@@ -148,6 +135,7 @@ int self_tests()
 	{
 		test_motor();
 	}
+#endif
 
     return test_bits;
 }
@@ -279,11 +267,13 @@ void enter_state(RespirationState new)
         sensors_stop_sampling_flow();
         //valve_inhale();
         VTe_start_mL = get_sensed_Vol_mL();
+		wait_ms(200); //TODO Find instant reverse
         motor_release();
 	}
 	else if(state==Exhalation) {
 		sensors_stop_sampling_flow();
         valve_exhale();
+		wait_ms(200);//TODO Find instant reverse
         motor_release();
 		VMe_Lpm = 0.f;
 		PEP_cmH2O = 0.f;
